backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
graphviz==0.20.1
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.1.4
pytz==2025.2
sqlparse==0.4.4
typing-extensions==4.7.1
